pyproject.toml
[project]
name = "jobbtest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django==6.0.2",
    "django-environ~=0.12.0",
    "django-ninja>=1.5.3",
    "postgis>=1.0.4",
    "psycopg~=3.3.2",
]
